% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_plots.R, R/plot_all.R
\name{example_plots}
\alias{example_plots}
\alias{example_plots}
\title{Example Plots}
\usage{
example_plots()

example_plots()
}
\description{
Plots four plots (hist, bar, dot, and box) in layout to test style.

Plots four plots (hist, bar, dot, and box) in layout to test style.
}
