% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPalette.R
\name{getPalette}
\alias{getPalette}
\title{Get Palette from CSV Files in Directory}
\usage{
getPalette(
  aesthetic,
  colors,
  colorOffset = 0,
  reverseOrder = FALSE,
  displayNames = FALSE,
  darkenPaletteForTextGeoms = FALSE,
  useColorBlindPalette = FALSE,
  colorBlindPaletteName = "safe",
  fileSuffix = "-colors",
  fileExtension = ".csv",
  directoryOfPalette = "extdata",
  thisPackageName = "styles"
)
}
\arguments{
\item{aesthetic}{The name of the aesthetic}

\item{colors}{A vector of color names}

\item{colorOffset}{Offsets the ordering of the palette. E.g., colorOffset = 2 would move the first two colors to the end of the palette, starting at the 3rd color.}

\item{reverseOrder}{Logical, TRUE reverses the default order of the palette. Default FALSE. Applied after colorOffset.}

\item{displayNames}{Logical, if TRUE display names of the colors}

\item{darkenPaletteForTextGeoms}{Logical, if TRUE then the palette will be darkened  slightly}

\item{useColorBlindPalette}{Logical, if TRUE uses color blind friendly palette instead of style.}

\item{colorBlindPaletteName}{Name of colorblind friendly palette. Use `cols4all::c4a_palettes()` for available palettes.}

\item{fileSuffix}{A suffix to be added to the file name (default: "-colors")}

\item{fileExtension}{The extension of the file (default: ".csv")}

\item{directoryOfPalette}{The directory of the palette data (default: "extdata")}

\item{thisPackageName}{The name of the package (default: "styles")}
}
\value{
The function returns a named list representing a color palette
}
\description{
This function reads palette data from CSV files located in a specified directory
}
\keyword{internal}
