% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimesNewRomanFamily.R
\name{getTimesNewRomanFamily}
\alias{getTimesNewRomanFamily}
\title{Get Times New Roman Family}
\usage{
getTimesNewRomanFamily()
}
\value{
The function returns a font family as a string.
}
\description{
Gets the times new roman family equivielent across systems
}
\examples{
getTimesNewRomanFamily()

library(ggplot2)
library(styles)

# Create basic ggplot object
ggplot(mtcars, aes(y=mpg, x=wt, color=as.factor(cyl))) +

  geom_point() +

  geom_text(aes(label = paste(mpg)),
            vjust = 1.2, hjust = -0.2,
            size = 3,

            # ** IMPORTANT: theme() does not control text labels
            # Need to set family to this font!
            family = styles::getTimesNewRomanFamily()
  ) +

  # This controls all non-label fonts
  theme_dc(inTimesNewRoman = TRUE)
}
