% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsetPalette.R
\name{offsetPalette}
\alias{offsetPalette}
\title{Offsets a palette by a number of colors}
\usage{
offsetPalette(colors, colorOffset)
}
\arguments{
\item{colors}{A vector of color names}

\item{colorOffset}{Offsets the ordering of the palette. E.g., colorOffset = 2 would move the first two colors to the end of the palette, starting at the 3rd color.}
}
\value{
The function returns a palette colors
}
\description{
This function offsets a palette by a number of colors
}
\keyword{internal}
