% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStyles.R
\name{scale_color_dbc}
\alias{scale_color_dbc}
\title{ggplot Color Scale}
\usage{
scale_color_dbc()
}
\value{
ggplot color scale
}
\description{
ggplot Color Scale
}
\examples{
library(ggplot2)
library(styles)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=wt, color=as.factor(cyl))) +
    geom_point(size = 3, alpha = 0.5)

# Add the color
ggplotObject + scale_color_dbc()
}
