% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStyles.R
\name{scale_dbc}
\alias{scale_dbc}
\title{Function that returns palettes or individual colors
This is done separately just to reduce noise in the file
Mainly, this just either displays the names of the colors or not when returing
It also allows you to return 1 or all of the colors}
\usage{
scale_dbc(aesthetic = NULL, ..., displayNames = FALSE)
}
\arguments{
\item{aesthetic}{Color aesthetic in 'fill', 'color', 'base', 'text','border','background', or 'grayedOut'}

\item{...}{single or vector of colors}

\item{displayNames}{TRUE if you want to see the names of each color. Default FALSE.}
}
\value{
hex codes for color palettes
}
\description{
Function that returns palettes or individual colors
This is done separately just to reduce noise in the file
Mainly, this just either displays the names of the colors or not when returing
It also allows you to return 1 or all of the colors
}
\examples{
library(styles)

# Single valued colors: Basic elements
scale_dbc('text')       # Text (dark gray)
scale_dbc('background') # Background (white)
scale_dbc('border')     # Border (light gray)
scale_dbc('grayedOut')  # (Very light gray)

# Palettes: Used in scale fill or scale color below
scale_dbc('color', displayNames = TRUE) # All colors
scale_dbc('fill',  displayNames = TRUE) # All fills
scale_dbc('fill', 3)                    # 3 fills
scale_dbc('fill', 'blue', displayNames = TRUE) # Just blue
}
