% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_dc.R
\name{scale_dc}
\alias{scale_dc}
\title{Function that returns a palette or a set of individual colors within the palette.}
\usage{
scale_dc(
  aesthetic = NULL,
  ...,
  colorOffset = 0,
  reverseOrder = FALSE,
  displayNames = FALSE,
  darkenPaletteForTextGeoms = FALSE,
  useColorBlindPalette = FALSE,
  colorBlindPaletteName = "safe"
)
}
\arguments{
\item{aesthetic}{Color aesthetic in 'fill', 'color', 'base', 'text','border','background', or 'grayedOut'}

\item{...}{single or vector of colors}

\item{colorOffset}{Offsets the ordering of the palette. E.g., colorOffset = 2 would move the first two colors to the end of the palette, starting at the 3rd color.}

\item{reverseOrder}{Logical, TRUE reverses the default order of the palette. Default FALSE. Applied after colorOffset.}

\item{displayNames}{TRUE if you want to see the names of each color. Default FALSE.}

\item{darkenPaletteForTextGeoms}{Logical, if TRUE then the palette ill be darkened slightly.}

\item{useColorBlindPalette}{Logical, if TRUE uses color blind friendly palette instead of style.}

\item{colorBlindPaletteName}{Name of colorblind friendly palette. Use `cols4all::c4a_palettes()` for available palettes.}
}
\value{
hex codes of a subsetted palette
}
\description{
Function that returns a palette or a set of individual colors within the palette.
}
\examples{
library(styles)

# Single valued colors: Basic elements
scale_dc('text') # Text (a dark gray)

# Palettes: Used in scale fill or scale color below
scale_dc('color', displayNames = TRUE) # All accent colors (line colors)
scale_dc('fill',  displayNames = TRUE) # All main colors (fills)
scale_dc('fill', 3)                    # 3 fills
scale_dc('fill', 'red', displayNames = TRUE) # Just a red fill

# Color blind friendly palette override
scale_dc('color', useColorBlindPalette = TRUE)

}
\seealso{
\code{\link{theme_dc}}

\code{\link{scale_color_dc}}

\code{\link{scale_fill_dc}}
}
