% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStyles.R
\name{scale_fill_dbc}
\alias{scale_fill_dbc}
\title{ggplot Fill Scale}
\usage{
scale_fill_dbc()
}
\value{
ggplot fill scale
}
\description{
ggplot Fill Scale
}
\examples{
library(ggplot2)
library(styles)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=wt, fill=as.factor(cyl))) +
    geom_point(size = 3, alpha = 0.5)

# Add the fill color
ggplotObject + scale_fill_dbc()
}
