% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_dc.R
\name{scale_fill_dc}
\alias{scale_fill_dc}
\title{ggplot Fill Scale}
\usage{
scale_fill_dc(
  overrideWithAccent = FALSE,
  colorOffset = 0,
  reverseOrder = FALSE,
  useColorBlindPalette = FALSE,
  colorBlindPaletteName = "safe",
  ...
)
}
\arguments{
\item{overrideWithAccent}{TRUE if you want to use the color palette for the fill mapping. (Default FALSE)}

\item{colorOffset}{Offsets the ordering of the palette. E.g., colorOffset = 2 would move the first two colors to the end of the palette, starting at the 3rd color.}

\item{reverseOrder}{Logical, TRUE reverses the default order of the palette. Default FALSE. Applied after colorOffset.}

\item{useColorBlindPalette}{Logical, if TRUE uses color blind friendly palette instead of style.}

\item{colorBlindPaletteName}{Name of colorblind friendly palette. Use `cols4all::c4a_palettes()` for available palettes.}

\item{...}{Additional parameters passed to either scale_color_manual or scale_fill_manual functions, such as guide = 'none'}
}
\value{
ggplot fill scale
}
\description{
ggplot Fill Scale
}
\examples{
library(ggplot2)
library(styles)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=as.factor(cyl), fill=as.factor(cyl))) +

    # Note that gray border used for demonstration. Please use scale_color_dc()
    # as much as possible
    geom_boxplot(size = 0.8, color = scale_dc('gray', 'gray3'))

# Add the fill color
ggplotObject + scale_fill_dc()

# Add the accent color
ggplotObject + scale_fill_dc(overrideWithAccent = TRUE)

# Add the fill color and turn off legend with optional parameter
ggplotObject + scale_fill_dc(guide = 'none')

# Offset the colors by 1
ggplotObject + scale_fill_dc(colorOffset = 1)

# reverse the order of the palette
ggplotObject + scale_fill_dc(reverseOrder = TRUE)

# Use color blind palette
ggplotObject + scale_fill_dc(useColorBlindPalette = TRUE)

}
\seealso{
\code{\link{theme_dc}}

\code{\link{scale_dc}}

\code{\link{scale_color_dc}}
}
