% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetAndFormatPalette.R
\name{subsetAndFormatPalette}
\alias{subsetAndFormatPalette}
\title{Format the palette vector based on the use input}
\usage{
subsetAndFormatPalette(colors, thePalette, displayNames)
}
\arguments{
\item{colors}{Vector of colors to include retrieve from the palette.}

\item{thePalette}{A palette to search on with the colors.}

\item{displayNames}{TRUE if you want to return a named list with the colors.}
}
\value{
The palette that only includes the subset of colors
}
\description{
Format the palette vector based on the use input
}
\keyword{internal}
