% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dc.R
\name{theme_dc}
\alias{theme_dc}
\title{Create a Theme for ggplot}
\usage{
theme_dc(borderMode = "default", ...)
}
\arguments{
\item{borderMode}{the theme borderMode, depending on `default` use, `facet` usage, or `borders`.}

\item{...}{Additional parameters that can be passed into the theme() function. Added last to theme.}
}
\value{
The function returns a ggplot theme
}
\description{
This function creates a custom theme for ggplot with several theme modifications
}
\examples{
library(ggplot2)
library(styles)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=wt, color=as.factor(cyl))) +
    geom_point(size = 3, alpha = 0.5) +
    scale_color_dc()

# Add the theme
ggplotObject + theme_dc()

# Add the theme with additional optional theme parameters
ggplotObject + theme_dc(legend.position = 'right')

}
\seealso{
\code{\link{scale_dc}}

\code{\link{scale_color_dc}}

\code{\link{scale_fill_dc}}
}
